activeNode moved to its own folder
