+++ src/inspection/activeNode/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import * as ActiveNodeUtils from "./activeNodeUtils";
+
+export { ActiveNodeUtils };
+export * from "./activeNode";
