+++ src/inspection/position/index.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
 import * as PositionUtils from "./positionUtils";
 
 export { PositionUtils };
